financial-package/financial-agents-service/pyproject.toml
[tool.poetry]
name = "financial-agents"
version = "0.1.0"
description = "Financial AI agents built with LangGraph"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ui"}, {include = "financial_agents"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
openai = ">=1.5.0"
langchain = ">=0.1.0"
langchain-core = ">=0.1.0"
langchain-community = ">=0.0.13"
langchain-openai = ">=0.0.2"
langgraph = ">=0.0.20"
python-dotenv = ">=1.0.0"
pydantic = ">=2.0.0"
chromadb = ">=0.4.18"
langchain-chroma = ">=0.1.0"
opensearch-py = {version = ">=2.0.0", optional = true}
pandas = ">=2.0.0"
numpy = ">=1.22.0"
sdv = ">=1.0.0"
tiktoken = ">=0.5.0"
tqdm = ">=4.65.0"
requests = ">=2.28.0"
nbformat = ">=5.10.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.1.0"
ipykernel = ">=6.0.0"
jupyter = ">=1.0.0"

[tool.poetry.group.ui.dependencies]
streamlit = ">=1.29.0"
fastapi = ">=0.103.0"
uvicorn = ">=0.23.0"

[tool.poetry.extras]
opensearch = ["opensearch-py"]
all = ["opensearch-py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 